+++ features/chat/klangk/lib/chat_message_list.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown_plus/flutter_markdown_plus.dart';
 import 'package:markdown/markdown.dart' as md;
-import '../theme/colors.dart';
-import '../utils/web_helpers_stub.dart'
-    if (dart.library.js_interop) '../utils/web_helpers_web.dart';
+import 'package:klangk_plugin_api/klangk_plugin_api.dart';
+import 'utils/open_url.dart';
 
 /// Scrollable list of chat messages with load-more and collapsible long
 /// messages.